+++ src/components/EditableLabel/styles.ts
@@ -1,14 +1,13 @@
-import styled, { css, keyframes } from 'styled-components';
+import styled, { keyframes } from 'styled-components';
 import CustomButton from '../CustomButton';
-import DefaultInput from '../Input';
 
 const animateInput = keyframes`
-0% {
-	opacity: 0;
-}
-100% {
-	opacity: 1;
-}
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
 `;
 
 export const Container = styled.div`
@@ -23,6 +22,10 @@ export const Container = styled.div`
     margin: 0;
     box-sizing: border-box;
     animation: ${animateInput} 0.5s ease-out;
+    
+    input {
+      width: 12rem;
+    }
   }
 `;
 
@@ -74,39 +77,6 @@ export const CancelButton = styled.button`
   border: none;
   background-color: transparent;
   cursor: pointer;
-`;
-
-export const Input = styled(DefaultInput)`
-  width: 12rem;
-  box-sizing: border-box;
-  > svg {
-    height: 1.0625rem;
-    width: 1.0625rem;
-    right: 0.6825rem;
-  }
-  > input {
-    box-sizing: border-box;
-    width: 100%;
-    margin: 0;
-    padding: 0.5rem;
-    border-radius: 0.75rem;
-    border: 1px solid #141414;
-    font-size: 0.6875rem;
-    font-weight: 700;
-    outline: none;
-    background-color: transparent;
-    box-shadow: none;
-    ${(props) =>
-      props.error
-        ? css`
-            border-color: #fb2e4c;
-            padding-right: 2.25rem;
-          `
-        : ''}
-    &:focus {
-      box-shadow: none;
-    }
-  }
 `;
 
 export const Overlay = styled.div`