+++ src/components/Icons/Alert/index.tsx
@@ -1,17 +1,16 @@
 import * as React from 'react';
 
-const Check: React.FC<React.SVGProps<SVGSVGElement>> = (props) => {
+const Alert: React.FC<React.SVGProps<SVGSVGElement>> = ({ fill = 'currentColor', ...props }) => {
   return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="104" height="104" viewBox="0 0 104 104" fill="none" {...props}>
-      <rect x="49" y="35" width="6" height="32" rx="3" fill={props.fill || '#AE8800'} />
-      <rect x="49" y="71" width="6" height="6" rx="3" fill={props.fill || '#AE8800'} />
+    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="11.5786" y="8.07715" width="1.5" height="7.38461" rx="0.75" fill={fill} />
+      <rect x="11.5786" y="16.3848" width="1.5" height="1.5" rx="0.75" fill={fill} />
       <path
-        d="M45.0718 14C48.151 8.66666 55.849 8.66667 58.9282 14L95.3013 77C98.3805 82.3333 94.5315 89 88.3731 89H15.6269C9.46853 89 5.61953 82.3333 8.69873 77L45.0718 14Z"
-        stroke={props.fill || '#AE8800'}
-        strokeWidth="3.5"
+        d="M10.5966 3.5C11.3664 2.16667 13.2909 2.16667 14.0607 3.5L22.0714 17.375C22.8412 18.7083 21.8789 20.375 20.3393 20.375H4.31788C2.77828 20.375 1.81603 18.7083 2.58583 17.375L10.5966 3.5Z"
+        stroke={fill}
       />
     </svg>
   );
 };
 
-export default Check;
+export default Alert;