+++ src/components/Icons/ViewMore/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-const ViewMore1: React.FC<React.SVGProps<SVGSVGElement>> = ({ fill = '#FB2E4C', ...props }) => {
+const ViewMore1: React.FC<React.SVGProps<SVGSVGElement>> = ({ fill = 'currentColor', ...props }) => {
   return (
     <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
       <g opacity="0.6">